Move IJsonRpcMessageBufferManager into Reflection namespace
Fixes #380
