+++ src/StreamJsonRpc/Reflection/IJsonRpcMessageBufferManager.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace StreamJsonRpc
+namespace StreamJsonRpc.Reflection
 {
     using StreamJsonRpc.Protocol;
